% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bunyas}
\alias{bunyas}
\title{Bunya Mountains tree data}
\format{
A list containing four data frames:
\describe{
  \item{abundances}{Abundance matrix of tree species recorded at five locations and subplots therein.Each row represents a 25 x 25 m subplot within a location. The columns (except for columns 1 and 2) are the names of rainforest tree species. The values within cells are the recorded abundance of each species in each subplot.}
  \item{sites}{Topographically-corrected climate data for the five locations extracted from CSIRO dat portal (Harwood et al. 2016).}
  \item{traits}{Functional trait data for most of the species. Values are averages taken from multiple individuals of each species. Traits include: ‘log_lamina_sla’ = log-transformed specific leaf area (original units mm2/mg-1), ‘log_lamina_area’ = log-transformed lamina (leaf) area (’lamina’ is the leaf blade excluding petioles and rachis; original units mm2), ‘lamina_ldmc’ = leaf dry matter content (dry mass over fresh mass), ‘tlp ‘= leaf turgor loss point (MPa), ‘wd’ = wood density (fresh volume / dry mass), ‘log_t_b’ = log(xylem wall thickness / xylem lumen breadth) (a proxy for conduit reinforcement), ‘sqrt_mh’ = sqrt-transformed maximum height (potential height of each tree species; original units m).}
  \item{trees}{Tree-scale data with diameter measurements. One row per individual tree, with species identity and a range of diameter measurements: ‘max_diam’ = the diameter of the largest stem. For most trees that have a single stem, this is just the diameter at breast height (1.3 m), ‘ba_m2’ = the basal area of each stem. For multi-stemmed individuals this is the sum of the cross-sectional area of all stems, ‘combined_diam’ = the diameter equivalent of the basal area. For single-stemmed trees this is the same as ‘max_diam’.}

  Dates: Tree survey data for ‘high’, ‘mid_east’ and ‘mid_west’ were collected by Don Butler in 2003. All trait data and tree survey data for ‘low_east’ and ‘low_west’ were collected by Alison Brown and John Dwyer in 2019.

  Location: Bunya Mountain National Park, Queensland Australia. Five areas of rainfortest or vine thicket were surveyed along a topographic moisture gradient:
'high' (easting = 359342, northing = 7027160)
'low_east' (365676, 7028551)
'low_west' (350888, 7030875)
'mid_east' (360196, 7027612)
'mid_west' (354971, 7031460)
All eastings and northings in WGS84, zone 56.
}
}
\source{
Brown, A. 2019. How does moisture stress structure angiosperm communities in subtropical rainforests? Honours Thesis. The University of Queensland, Brisbane.

Harwood, T., R. Donohue, I. Harman, T. McVicar, N. Ota, J. Perry, and K. Williams. 2016. 9s climatology for continental Australia 1976-2005: Summary variables with elevation and radiative adjustment v3. CSIRO Data Collection.
}
\usage{
bunyas
}
\description{
Tree community dataset from rainforests and vine thickets of the Bunya Mountains, QLD, Australia.
}
\author{
Alison Brown \email{alison.brown25@gmail.com}

Don Butler \email{don.butler@des.qld.gov.au}

John Dwyer \email{j.dwyer2@uq.edu.au}
}
\keyword{datasets}
