% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drosophilaWingG}
\alias{drosophilaWingG}
\title{Drosophila G matrix}
\format{
A symmetrical 10x10 matrix where elements on the diagonal are estimated variances and off-diagonal elements are covariances in ten interlandmark distances between nine wing landmarks (see Fig. 1 in McGuigan & Blows 2007).
}
\source{
McGuigan, K., Blows, M.W. The phenotypic and genetic covariance structure of Drosophilid wings. Evolution 61: 902–911 (2007)
}
\usage{
drosophilaWingG
}
\description{
A matrix containing estimated genetic variances and covariances in morphological wing traits of female Drosophila bunnanda flies.
}
\keyword{datasets}
