% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nectarYeast}
\alias{nectarYeast}
\title{Nectar yeast data}
\format{
A data frame with 12 columns:
\describe{
  \item{strain}{Focal species of yeast, including Metschnikowia reukaufii (Mr), M. koreensis (Mk), M. gruessii (Mg) and Starmerella bombicola (Sb)}
  \item{rep}{Replicate experiment}
  \item{exp}{Experiment ID (one of two batches)}
  \item{treat.env}{Treatment environment, either constant (constant sucrose levels) or variable (fluctuating sucrose levels)}
  \item{treat.bio}{Monoculture (mono) or competition (comp) between two species}
  \item{mixID}{For competition experiments, the species identities of the two competing species}
  \item{AA.mM}{Amino acid concentration in mM}
  \item{sucrose}{Sucrose levels, osmotic pressure in percent}
  \item{timepoint}{Time in hours}
  \item{CFUs}{Number of yeast cells ("colony-forming units") per uL}
  \item{validity}{Should data be included in analysis?}
  \item{notes}{Notes}
}
}
\source{
Letten, A.D. Dhami, M.K., Ke, P.J., T Fukami, T. Species coexistence through simultaneous fluctuation-dependent mechanisms. Proceedings of the National Academy of Sciences 115 (26), 6745-6750 (2018).
\url{https://www.pnas.org/content/115/26/6745.full}
}
\usage{
nectarYeast
}
\description{
Data from an experiment investigating growth of different nectar yeasts in both monoculture and competition.
}
\keyword{datasets}
