% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxley}
\alias{oxley}
\title{Oxley seedling survival data}
\format{
A list containing four data frames:
\describe{
  \item{trees}{}
  \item{soils}{}
  \item{plots}{}
  \item{species}{}
}
}
\source{
Gardiner, R., Shoo, L.P., Dwyer, J.M. Look to seedling heights, rather than functional traits, to explain survival during extreme heat stress in the early stages of subtropical rainforest restoration. Journal of Applied Ecology 56:2687-2697 (2019).
DOI: 10.1111/1365-2664.13505, \url{https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/1365-2664.13505}
}
\usage{
oxley
}
\description{
Data from an experiment investigating tree seedling survival and growth to inform subtropical rainforest restoration efforts
}
\keyword{datasets}
