% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rainbowfishG}
\alias{rainbowfishG}
\title{Rainbowfish G matrix}
\format{
A symmetrical 6x6 matrix where elements on the diagonal are estimated variances (x1000) and off-diagonal elements are covariances (x1000) in the following traits:
\describe{
  \item{bodyL}{Standard length}
  \item{predorsL}{Predorsal length}
  \item{headD}{Head depth}
  \item{bodyD}{Body depth}
  \item{caudPedL}{Caudal peduncle length}
  \item{caudPedD}{Caudal peduncle depth}
}
}
\source{
McGuigan, K., Chenoweth, S.F., Blows, M.W. Phenotypic divergence along lines of genetic variance. American Naturalist 165: 32–43 (2005)

and

McGuigan, K. Studying phenotypic evolution using multivariate quantitative genetics. Molecular Ecology 15: 883-896 (2006)
}
\usage{
rainbowfishG
}
\description{
A matrix containing variances and covariances in morphological traits of the Lake Eacham rainbow fish (Melanotaenia eachamensis).
}
\keyword{datasets}
