% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rainforestFire}
\alias{rainforestFire}
\title{Rainforest fire and mammal diversity}
\format{
A list containing two tibbles:
\describe{
  \item{images}{Table of camera images taken at different sites, and with identified animals for each picture.}
  \item{sites}{Table of sites, including geographical location, fire habitat category, and various other information.}
  }
}
\source{
Bird, R.R., Zsoldos, R.R., Jimenez Sandoval, M.V., Watson, S.J. & Smith, A.L. Wildfire in rainforest margins is associated with variation in mammal diversity and habitat use. (under review)
}
\usage{
rainforestFire
}
\description{
Camera trap data to on mammal communities in rainforest margins that are affected to varying degrees by wildfires.
}
\keyword{datasets}
