% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reefFishFST}
\alias{reefFishFST}
\title{Reef fish FST data}
\format{
A data frame with 114 rows and 12 columns. Each row corresponds to a different species and the most important variables are:
\describe{
  \item{Reference}{Reference number to original article (see source)}
  \item{Fam_Spp}{String containing family and species names}
  \item{Egg}{Reproductive strategy with respect to eggs}
  \item{Marker}{Type of genetic marker used in the study}
  \item{FST}{FST estimate}
}
}
\source{
Riginos, C., Buckley, Y.M., Blomberg, S.P. & Treml, E.A. 2014.
The American Naturalist 184: 52-64 (2014)
}
\usage{
reefFishFST
}
\description{
A dataset containing published estimates of FST values for 114 species of reef fish.
}
\keyword{datasets}
