% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reefFishPhylogeny}
\alias{reefFishPhylogeny}
\title{Reef fish phylogenetic tree}
\format{
Phylogenetic tree as an object of class phylo, as specified in the ape package. When ape is loaded, the object will be displayed and handled as a phylogenetic tree. When ape is not loaded, this is simply a list with the following elements:
\describe{
  \item{edge}{matrix where each row corresponds to an edge (branch) of the tree and the two columns identify the node numbers that the branch connects}
  \item{edge.length}{Branch lengths in million years}
  \item{Nnode}{Number of nodes in the tree}
  \item{node.label}{Labels for some of the internal nodes in the tree}
  \item{tip.label}{Species names at the tips of the tree}
}
}
\source{
Riginos, C., Buckley, Y.M., Blomberg, S.P. & Treml, E.A. 2014.
The American Naturalist 184: 52-64 (2014)
}
\usage{
reefFishPhylogeny
}
\description{
A dataset specifying a dated phylogenetic tree of 154 species of reef fish.
}
\keyword{datasets}
