% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senecio}
\alias{senecio}
\title{Senecio lautus morphological data}
\format{
A data frame with the following columns:
\describe{
  \item{Ecotype}{One of four ecotypes (dune, headland, tableland woodland)}
  \item{Population}{Population ID}
  \item{VegHeight}{Vegetative height of the plant, in mm}
  \item{MSL_W}{Ratio between main stem length and mean plant width}
  \item{SB}{Number of branches}
  \item{MSD}{Main stem diameter, in mm}
  \item{Area}{Leaf area, in mm2}
  \item{P2A2}{Leaf perimeter squared / area squared (an indicator of leaf complexity)}
  \item{Circularity}{Leaf circularity}
  \item{Nindents.Peri}{Number of leaf indents divided by leaf perimeter}
  \item{IndentWidth}{Leaf indent width, in mm}
  \item{IndentDepth}{Leaf ndent depth, in mm}
}
}
\source{
Walter, G.M., Aguirre, J.D., Blows, M.W. & Ortiz-Barrientos, D. Evolution of Genetic Variance during Adaptive Radiation. The American Naturalist 191: E108–E128 (2018), \url{https://www.journals.uchicago.edu/doi/10.1086/696123}
}
\usage{
senecio
}
\description{
Data from an experiment where seeds from the Australian wildflower \emph{Senecio lautus} were collected, grown in glasshouses under uniform conditions, and a range of morphological traits measured. The seeds were collected from populations belonging to four distinct ecotypes (dune, headland, tableland and woodland) that may be in the process of forming separate species.
}
\keyword{datasets}
